refactor! Rename operation classes to `xxxHandler`
Adds a `Handler` suffix to all current final operation handler class names, and
moves them to a `handlers` namespace.

This is not expected to affect end-users as the in-built handler classes are not
generally expected to be used outside the package.

This will avoid naming conflicts and confusion when we subsequently introduce
support for representing a patch as a set of operation DTOs.