feat: Support building & using patches as a list of DTOs
Build on the existing implementation to allow users to specify patches
as a list of operation objects, rather than as a JSON string.

Unlike the JSON form, the patch operation DTOs are strict about the
parameters they accept. If a user wishes to include custom properties,
they can implement a class extending the base `PatchOperation`.

Patch DTOs can be serialised to and from JSON, for convenience.