feat: Make all `final` patch operation classes `readonly`
Will break on 8.1, to be rebased once the project drops 8.1 support.

I have intentionally left the base `PatchOperation` class as *not*
readonly in case end-users want to extend it with mutable operation
DTOs for any reason. Instead, I have just left `op` as a readonly
property in the base class.