test: Prove that PatchOperationList::fromJson bubbles unexpected errors
It's essentially impossible for our own DTOs to trigger PHP errors other
than the ones related to creating them with incorrect parameters.

However, because our code has to catch `Error` to check for certain
types of error, we need to prove that it rethrows if the error is not
one that it expects to handle.